+++ impyute/imputation/cs/central_tendency.py
@@ -1,10 +1,9 @@
 import numpy as np
-from impyute.ops import find_null
-from impyute.ops import checks
-from impyute.ops import preprocess
+from impyute.ops import matrix
+from impyute.ops import wrapper
 
-@preprocess
-@checks
+@wrapper.wrappers
+@wrapper.checks
 def mean(data):
     """ Substitute missing values with the mean of that column.
 
@@ -19,15 +18,15 @@ def mean(data):
         Imputed data.
 
     """
-    null_xy = find_null(data)
-    for x_i, y_i in null_xy:
+    nan_xy = matrix.nan_indices(data)
+    for x_i, y_i in nan_xy:
         row_wo_nan = data[:, [y_i]][~np.isnan(data[:, [y_i]])]
         new_value = np.mean(row_wo_nan)
         data[x_i][y_i] = new_value
     return data
 
-@preprocess
-@checks
+@wrapper.wrappers
+@wrapper.checks
 def median(data):
     """ Substitute missing values with the median of that column(middle).
 
@@ -42,19 +41,19 @@ def median(data):
         Imputed data.
 
     """
-    null_xy = find_null(data)
-    cols_missing = set(null_xy.T[1])
+    nan_xy = matrix.nan_indices(data)
+    cols_missing = set(nan_xy.T[1])
     medians = {}
     for y_i in cols_missing:
         cols_wo_nan = data[:, [y_i]][~np.isnan(data[:, [y_i]])]
         median_y = np.median(cols_wo_nan)
         medians[str(y_i)] = median_y
-    for x_i, y_i in null_xy:
+    for x_i, y_i in nan_xy:
         data[x_i][y_i] = medians[str(y_i)]
     return data
 
-@preprocess
-@checks
+@wrapper.wrappers
+@wrapper.checks
 def mode(data):
     """ Substitute missing values with the mode of that column(most frequent).
 
@@ -72,7 +71,7 @@ def mode(data):
         Imputed data.
 
     """
-    null_xy = find_null(data)
+    nan_xy = matrix.nan_indices(data)
     modes = []
     for y_i in range(np.shape(data)[1]):
         unique_counts = np.unique(data[:, [y_i]], return_counts=True)
@@ -80,6 +79,6 @@ def mode(data):
         mode_y = [unique for unique, count in np.transpose(unique_counts)
                   if count == max_count and not np.isnan(unique)]
         modes.append(mode_y)  # Appends index of column and column modes
-    for x_i, y_i in null_xy:
+    for x_i, y_i in nan_xy:
         data[x_i][y_i] = np.random.choice(modes[y_i])
     return data