+++ impyute/imputation/cs/random.py
@@ -1,10 +1,9 @@
 import numpy as np
-from impyute.ops import find_null
-from impyute.ops import preprocess
-from impyute.ops import checks
+from impyute.ops import matrix
+from impyute.ops import wrapper
 
-@preprocess
-@checks
+@wrapper.wrappers
+@wrapper.checks
 def random(data):
     """ Fill missing values in with a randomly selected value from the same
     column.
@@ -20,8 +19,8 @@ def random(data):
         Imputed data.
 
     """
-    null_xy = find_null(data)
-    for x, y in null_xy:
+    nan_xy = matrix.nan_indices(data)
+    for x, y in nan_xy:
         uniques = np.unique(data[:, y])
         uniques = uniques[~np.isnan(uniques)]
         data[x][y] = np.random.choice(uniques)