+++ impyute/ops/__init__.py
@@ -1,13 +1,12 @@
 """ Unorganized set of utility functions """
 
-from .matrix import find_null, map_nd, every_nd
-from .errors import BadInputError, BadOutputError
-from .wrapper import checks
-from .wrapper import preprocess
-from . import wrapper
+from . import error
 from . import inverse_distance_weighting
+from . import matrix
 from . import util
+from . import wrapper
 
-__all__ = ["find_null", "map_nd", "every_nd", "checks",
-           "BadInputError", "BadOutputError", "preprocess",
-           "inverse_distance_weighting", "util", "wrapper"]
+__all__ = [
+        "error", "inverse_distance_weighting", "matrix",
+        "util", "wrapper"
+        ]