+++ impyute/ops/inverse_distance_weighting.py
@@ -1,5 +1,4 @@
 """ Assign weights to distances in a way such that farther values are weighed less """
-
 import numpy as np
 
 def shepards(distances, power=2):