+++ test/dataset/test_mnist.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 from impyute.dataset import mnist
-from impyute.ops import find_null
+from impyute.ops import matrix
 
 pytest.skip("takes ~30 sec each test", allow_module_level=True)
 data = mnist()["X"]
@@ -12,4 +12,4 @@ def test_return_type():
 
 def test_missing_values_present():
     """ Check that the dataset is corrupted (missing values present)"""
-    assert find_null(data).size != 0
+    assert matrix.nan_indices(data).size != 0