+++ test/ops/test_wrapper.py
@@ -1,10 +1,7 @@
 import pytest
 import numpy as np
 from impyute.imputation.cs import mean
-from impyute.ops import preprocess
-from impyute.ops import BadInputError
-from impyute.ops import BadOutputError
-from impyute.ops import checks
+from impyute.ops import error
 from impyute.ops import wrapper
 
 # pylint:disable=redefined-builtin
@@ -17,27 +14,27 @@ except NameError:
 except ModuleNotFoundError:
     pass
 
-@preprocess
-def preprocess_mul(arr):
+@wrapper.wrappers
+def wrappers_mul(arr):
     """Some function that performs an inplace operation on the input. Accepts kwargs"""
     arr *= 25
     return arr
 
-def test_preprocess_inplace_false():
+def test_wrappers_inplace_false():
     """Input should be unchanged if inplace set to false"""
     A = np.ones((5, 5))
     A_copy = A.copy()
-    preprocess_mul(A, inplace=False)
+    wrappers_mul(A, inplace=False)
     assert A[0, 0] == A_copy[0, 0]
 
-def test_preprocess_inplace_true():
+def test_wrappers_inplace_true():
     """Input may be changed if inplace set to true and operation is inplace"""
     A = np.ones((5, 5))
     A_copy = A.copy()
-    preprocess_mul(A, inplace=True)
+    wrappers_mul(A, inplace=True)
     assert A[0, 0] != A_copy[0, 0]
 
-def test_preprocess_pandas_input():
+def test_wrappers_pandas_input():
     """ Input: DataFrame, Output: DataFrame """
     # Skip this test if you don't have pandas
     pytest.importorskip('pandas')
@@ -49,7 +46,7 @@ def test_preprocess_pandas_input():
     # Assert that the output is a DataFrame
     assert isinstance(mean(A), pd.DataFrame)
 
-@checks
+@wrapper.checks
 def some_fn(data):
     """Dummy fn that has form of np.array -> np.array"""
     return data
@@ -62,25 +59,25 @@ def test_correct_input():
     arr.dtype = np.float
     try:
         some_fn(arr)
-    except BadInputError:
+    except error.BadInputError:
         assert False
 
 def test_1d():
     """ Check 1d array, BadInputError raised"""
     arr = np.array([np.nan, 2])
-    with pytest.raises(BadInputError) as excinfo:
+    with pytest.raises(error.BadInputError) as excinfo:
         some_fn(arr)
     assert str(excinfo.value) == "No support for arrays that aren't 2D yet."
 
 def test_not_nparray():
     """ If not an np.array, BadInputError raised"""
-    with pytest.raises(BadInputError) as excinfo:
+    with pytest.raises(error.BadInputError) as excinfo:
         some_fn([[np.nan, 2.], [3, 4]])
     assert str(excinfo.value) == "Not a np.ndarray."
 
 def test_nan_exists():
     """ If no NaN, BadInputError raised"""
-    with pytest.raises(BadInputError) as excinfo:
+    with pytest.raises(error.BadInputError) as excinfo:
         some_fn(np.array([[1.]]))
     assert str(excinfo.value) == "No NaN's in given data"
 
@@ -130,6 +127,6 @@ def test_conform_output_valid():
     encountered. First invalid value is 1.1
     """
     arr = np.array([[1.1, 0.5], [0.2, -1]])
-    with pytest.raises(BadOutputError) as excinfo:
+    with pytest.raises(error.BadOutputError) as excinfo:
         conform_output_dummy(arr, valid_fn=is_between_0_1)
     assert str(excinfo.value) == "1.1 does not conform"